Separate project status loggers
